+++ frontend/src/components/pages/CreateGame/components/GameMode.js
@@ -3,11 +3,13 @@ import React from "react";
 const GameMode = ({gameMode, pickGameMode, color}) => {
    const {label, description, allowed} = gameMode;
     return (
+        <div className="game-mode-wrapper">
         <button className={`game-mode ${color}`} onClick={() => pickGameMode(gameMode)} disabled={!allowed}>
             <p className="game-mode-label">{label}</p>
             <p className="game-mode-description">{description}</p>
         </button>
+        </div>
     );
-}
+};
 
 export default GameMode;
